+++ src/main/java/org/openpnp/machine/reference/ReferenceNozzleTip.java
@@ -3,10 +3,12 @@ package org.openpnp.machine.reference;
 import java.awt.event.ActionEvent;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Locale;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import javax.swing.AbstractAction;
 import javax.swing.Action;
@@ -14,13 +16,14 @@ import javax.swing.JOptionPane;
 
 import org.apache.commons.io.IOUtils;
 import org.opencv.core.KeyPoint;
-import org.opencv.core.RotatedRect;
 import org.openpnp.ConfigurationListener;
 import org.openpnp.gui.MainFrame;
 import org.openpnp.gui.support.Icons;
 import org.openpnp.gui.support.PropertySheetWizardAdapter;
 import org.openpnp.gui.support.Wizard;
+import org.openpnp.machine.reference.wizards.ReferenceNozzleTipCalibrationWizard;
 import org.openpnp.machine.reference.wizards.ReferenceNozzleTipConfigurationWizard;
+import org.openpnp.model.AbstractModelObject;
 import org.openpnp.model.Configuration;
 import org.openpnp.model.LengthUnit;
 import org.openpnp.model.Location;
@@ -37,6 +40,7 @@ import org.openpnp.util.UiUtils;
 import org.openpnp.util.VisionUtils;
 import org.openpnp.vision.pipeline.CvPipeline;
 import org.openpnp.vision.pipeline.CvStage.Result;
+import org.pmw.tinylog.Logger;
 import org.simpleframework.xml.Attribute;
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.ElementList;
@@ -192,7 +196,10 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
 
     @Override
     public PropertySheet[] getPropertySheets() {
-        return new PropertySheet[] {new PropertySheetWizardAdapter(getConfigurationWizard())};
+        return new PropertySheet[] {
+                new PropertySheetWizardAdapter(getConfigurationWizard()),
+                new PropertySheetWizardAdapter(new ReferenceNozzleTipCalibrationWizard(this), "Calibration")
+                };
     }
 
     public boolean isAllowIncompatiblePackages() {
@@ -275,7 +282,7 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
         this.changerMid2ToEndSpeed = changerMid2ToEndSpeed;
     }
 
-    private Nozzle getParentNozzle() {
+    public Nozzle getParentNozzle() {
         for (Head head : Configuration.get().getMachine().getHeads()) {
             for (Nozzle nozzle : head.getNozzles()) {
                 for (NozzleTip nozzleTip : nozzle.getNozzleTips()) {
@@ -306,6 +313,14 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
 
     public Calibration getCalibration() {
         return calibration;
+    }
+    
+    public void calibrate() throws Exception {
+        getCalibration().calibrate(this);
+    }
+    
+    public boolean isCalibrated() {
+        return getCalibration().isCalibrated();
     }
 
     public Action loadAction = new AbstractAction("Load") {
@@ -357,109 +372,401 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
     };
 
     @Root
-    public static class Calibration {
-        public static class CalibrationOffset {
-            final Location offset;
-            final double angle;
+    public static class Calibration extends AbstractModelObject {
+        public static interface RunoutCompensation {
 
-            public CalibrationOffset(Location offset, double angle) {
-                this.offset = offset;
-                this.angle = angle;
+            Location getOffset(double angle);
+            
+            Location getAxisOffset();
+            
+            @Override
+            String toString();
+        }
+        
+        public static class TableBasedRunoutCompensation implements RunoutCompensation {
+            List<Location> nozzleTipMeasuredLocations;
+
+
+            public TableBasedRunoutCompensation(List<Location> nozzleTipMeasuredLocations) {
+                //store data for later usage
+                this.nozzleTipMeasuredLocations = nozzleTipMeasuredLocations;
+            }
+
+            public Location getOffset(double angle) {
+                
+                // find the two angles in the measurements, that angle is between
+                List<Location> offsets = getOffsetPairForAngle(angle);
+                
+                // the xy-offset is available via getX/getY. the angle is available via getRotation() - it's stored this way because then a simple Location type is sufficient
+                Location offsetA = offsets.get(0);
+                Location offsetB = offsets.get(1).convertToUnits(offsetA.getUnits());    // could this conversion be omitted?
+
+                double ratio = 1.0;     // TODO Better solution than the workaround seems to be recommended.
+                if ( (offsetB.getRotation() - offsetA.getRotation()) != 0 ) {
+                    ratio = (angle - offsetA.getRotation()) / (offsetB.getRotation() - offsetA.getRotation());
+                }
+                double deltaX = offsetB.getX() - offsetA.getX();
+                double deltaY = offsetB.getY() - offsetA.getY();
+                double offsetX = offsetA.getX() + (deltaX * ratio);
+                double offsetY = offsetA.getY() + (deltaY * ratio);
+
+                return new Location(offsetA.getUnits(), offsetX, offsetY, 0, 0);
+            }
+            
+            /**
+             * Find the two closest offsets to the angle being requested. The offsets start at first measurement at angleStart
+             * and go to angleStop
+             */
+            private List<Location> getOffsetPairForAngle(double angle) {
+                Location a = null, b = null;
+                
+                // angle asked for is the last in the table?
+                if (angle >= nozzleTipMeasuredLocations.get(nozzleTipMeasuredLocations.size() - 1).getRotation()) {
+                    return Arrays.asList(nozzleTipMeasuredLocations.get(nozzleTipMeasuredLocations.size() - 1), nozzleTipMeasuredLocations.get(0));
+                }
+                
+                for (int i = 0; i < nozzleTipMeasuredLocations.size(); i++) {
+                    if (angle < nozzleTipMeasuredLocations.get(i + 1).getRotation()) {
+                        a = nozzleTipMeasuredLocations.get(i);
+                        b = nozzleTipMeasuredLocations.get(i + 1);
+                        break;
+                    }
+                }
+                return Arrays.asList(a, b);
             }
 
             @Override
             public String toString() {
-                return angle + " " + offset;
+                return String.format(Locale.US, "0° Offset %f, %f", nozzleTipMeasuredLocations.get(0).getX(), nozzleTipMeasuredLocations.get(0).getY());
+            }
+
+            @Override
+            public Location getAxisOffset() {
+                // axis offset is not available with this algorithm
+                return null;
             }
         }
 
+        public static class ModelBasedRunoutCompensation implements RunoutCompensation {
+        	List<Location> nozzleTipMeasuredLocations;
+        	
+            double centerX;
+            double centerY;
+            double radius;
+            
+            double phaseShift;
+
+			public ModelBasedRunoutCompensation(List<Location> nozzleTipMeasuredLocations) {
+				//store data for possible later usage
+				this.nozzleTipMeasuredLocations = nozzleTipMeasuredLocations;
+				
+				// first calculate the circle fit and store the values to centerXY and radius
+				// the measured offsets describe a circle with the rotational axis as the center, the runout is the circle radius
+				this.calcCircleFitKasa(nozzleTipMeasuredLocations);
+				
+				// afterwards calc the phaseShift angle mapping
+				this.calcPhaseShift(nozzleTipMeasuredLocations);
+            }
+			
+			/* function to calc the model based offset in cartesian coordinates */
+			public Location getOffset(double angle) {
+
+	            //add phase shift
+	            angle = angle - this.phaseShift;
+	            
+	            angle = Math.toRadians(angle);
+	            
+	            // convert from polar coords to xy cartesian offset values
+	            double offsetX = this.centerX + (this.radius * Math.cos(angle));
+	            double offsetY = this.centerY + (this.radius * Math.sin(angle));
+
+	            return new Location(LengthUnit.Millimeters, offsetX, offsetY, 0, 0);
+			}
+			
+	        private void calcCircleFitKasa(List<Location> nozzleTipMeasuredLocations) {
+	        	/* 
+	        	 * this function fits a circle my means of the Kasa Method to the given List<Location>.
+	        	 * this is a java port of http://people.cas.uab.edu/~mosya/cl/CPPcircle.html 
+	        	 * The Kasa method should work well for this purpose since the measured locations are captured along a full circle
+	        	 */
+	    		int n;
+
+	    	    double kasaXi,kasaYi,kasaZi;
+	    	    double kasaMxy,kasaMxx,kasaMyy,kasaMxz,kasaMyz;
+	    	    double kasaB,kasaC,kasaG11,kasaG12,kasaG22,kasaD1,kasaD2;
+	    	    double kasaMeanX=0.0, kasaMeanY=0.0;
+	    	    
+	    	    n = nozzleTipMeasuredLocations.size();
+
+	            Iterator<Location> nozzleTipMeasuredLocationsIterator = nozzleTipMeasuredLocations.iterator();
+	    		while (nozzleTipMeasuredLocationsIterator.hasNext()) {
+	    			Location measuredLocation = nozzleTipMeasuredLocationsIterator.next();
+	    			kasaMeanX += measuredLocation.getX();
+	    			kasaMeanY += measuredLocation.getY();
+	    		}
+	    		kasaMeanX = kasaMeanX / (double)nozzleTipMeasuredLocations.size();
+	    		kasaMeanY = kasaMeanY / (double)nozzleTipMeasuredLocations.size();
+	    		
+	    	    kasaMxx=kasaMyy=kasaMxy=kasaMxz=kasaMyz=0.;
+
+	    	    for (int i = 0; i < n; i++) {
+	    	        kasaXi = nozzleTipMeasuredLocations.get(i).getX() - kasaMeanX;   //  centered x-coordinates
+	    	        kasaYi = nozzleTipMeasuredLocations.get(i).getY() - kasaMeanY;   //  centered y-coordinates
+	    	        kasaZi = kasaXi*kasaXi + kasaYi*kasaYi;
+
+	    	        kasaMxx += kasaXi*kasaXi;
+	    	        kasaMyy += kasaYi*kasaYi;
+	    	        kasaMxy += kasaXi*kasaYi;
+	    	        kasaMxz += kasaXi*kasaZi;
+	    	        kasaMyz += kasaYi*kasaZi;
+	    	    }
+	    	    kasaMxx /= n;
+	    	    kasaMyy /= n;
+	    	    kasaMxy /= n;
+	    	    kasaMxz /= n;
+	    	    kasaMyz /= n;
+
+		    	// solving system of equations by Cholesky factorization
+	    	    kasaG11 = Math.sqrt(kasaMxx);
+	    	    kasaG12 = kasaMxy / kasaG11;
+	    	    kasaG22 = Math.sqrt(kasaMyy - kasaG12 * kasaG12);
+
+	    	    kasaD1 = kasaMxz / kasaG11;
+	    	    kasaD2 = (kasaMyz - kasaD1*kasaG12)/kasaG22;
+
+	    	    // computing parameters of the fitting circle
+	    	    kasaC = kasaD2/kasaG22/2.0;
+	    	    kasaB = (kasaD1 - kasaG12*kasaC)/kasaG11/2.0;
+	    	    
+	    	    // assembling the output
+	    	    this.centerX = kasaB + kasaMeanX;
+                this.centerY = kasaC + kasaMeanY;
+                this.radius = Math.sqrt(kasaB*kasaB + kasaC*kasaC + kasaMxx + kasaMyy);
+                
+                Logger.debug("[nozzleTipCalibration]calculated nozzleEccentricity: {}", this.toString());
+	        }
+
+            private void calcPhaseShift(List<Location> nozzleTipMeasuredLocations) {
+            	/*
+            	 * The phaseShift is calculated to map the angle the nozzle is located mechanically at
+            	 * (that is what openpnp shows in the DRO) to the angle, the nozzle tip is located wrt. to the
+            	 * centered circle runout path.
+            	 * With the phaseShift available, the calibration offset can be calculated analytically for every
+            	 * location/rotation even if not captured while measured (stepped by angleIncrement)
+            	 * 
+            	 */
+            	double phaseShift = 0;
+            	
+            	double angle=0;
+            	double measuredAngle=0;
+            	double differenceAngleMean=0;
+            	
+                Iterator<Location> nozzleTipMeasuredLocationsIterator = nozzleTipMeasuredLocations.iterator();
+        		while (nozzleTipMeasuredLocationsIterator.hasNext()) {
+        			Location measuredLocation = nozzleTipMeasuredLocationsIterator.next();
+        			
+        			// get the measurement rotation
+            		angle = measuredLocation.getRotation();		// the angle at which the measurement was made was stored to the nozzleTipMeasuredLocation into the rotation attribute
+            		
+            		// move the offset-location by the centerY/centerY. by this all offset-locations are wrt. the 0/0 origin
+            		Location centeredLocation = measuredLocation.subtract(new Location(LengthUnit.Millimeters,this.centerX,this.centerY,0.,0.));
+            		
+            		// calculate the angle, the nozzle tip is located at
+            		measuredAngle=Math.toDegrees(Math.atan2(centeredLocation.getY(), centeredLocation.getX()));
+            		
+            		// the difference is the phaseShift
+            		double differenceAngle = angle-measuredAngle;
+            		
+            		// atan2 outputs angles from -PI to +PI. If one wants positive values, one needs to add +PI to negative values
+            		if(differenceAngle < -180) {
+            			differenceAngle += 360;
+            		}
+            		if(differenceAngle > 180) {
+            			// since calculating the difference angle in some circumstances the angle can be bigger than 360 -> subtract
+            			differenceAngle -= 360;
+            		}
+            		
+            		Logger.trace("[nozzleTipCalibration]differenceAngle: {}", differenceAngle);
+            		
+            		// sum up all differenceAngles to build the average later
+            		differenceAngleMean += differenceAngle;
+        		}
+            	
+        		// calc the average
+            	phaseShift = differenceAngleMean / nozzleTipMeasuredLocations.size();
+                
+            	this.phaseShift = phaseShift;
+            	
+            	Logger.debug("[nozzleTipCalibration]calculated phaseShift: {}", this.phaseShift);
+            }
+            
+            
+            @Override
+            public String toString() {
+                return String.format(Locale.US, "Center %f, %f, Runout %f", centerX, centerY, radius);
+            }
+
+            @Override
+            public Location getAxisOffset() {
+                return new Location(LengthUnit.Millimeters,centerX,centerY,0.,0.);
+            }
+        }
+        
         @Element(required = false)
         private CvPipeline pipeline = createDefaultPipeline();
 
         @Attribute(required = false)
-        private double angleIncrement = 15;
+        private int angleSubdivisions = 6;
+        @Attribute(required = false)
+        private double angleStart = -180;
+        @Attribute(required = false)
+        private double angleStop = 180;
         
         @Attribute(required = false)
         private boolean enabled;
         
         private boolean calibrating;
+        
+        private RunoutCompensation runoutCompensation = null;
+        
+        public enum RunoutCompensationAlgorithm {
+            Model, Table
+        }
+        
+        @Attribute(required = false)
+        private RunoutCompensationAlgorithm runoutCompensationAlgorithm = RunoutCompensationAlgorithm.Model;      // modelBased or tableBased? Two implementations are available
 
-        List<CalibrationOffset> offsets;
+        public RunoutCompensationAlgorithm getRunoutCompensationAlgorithm() {
+            return this.runoutCompensationAlgorithm;
+        }
 
+        public void setRunoutCompensationAlgorithm(RunoutCompensationAlgorithm runoutCompensationAlgorithm) {
+            this.runoutCompensationAlgorithm = runoutCompensationAlgorithm;
+        }
+
+        public String getRunoutCompensationInformation() {
+            if(isCalibrated()) {
+                return runoutCompensation.toString();
+            } else {
+                return "Uncalibrated";
+            }
+        }
+        
         public void calibrate(ReferenceNozzleTip nozzleTip) throws Exception {
-            if (!isEnabled()) {
+        	/* TODO, possible later refinement:
+        	 * a) add plausibility checks
+        	 * b) if one measurement fails, one could retry or skip that silently if enough valid measurements are still available (possible only for modelBased algorithm)
+        	 * c) polish the gui 
+        	 */
+        	
+            if ( !isEnabled() ) {
                 return;
             }
+            
+            if ( !Configuration.get().getMachine().isHomed() ) {
+                Logger.trace("[nozzleTipCalibration]Machine not yet homed, nozzle tip calibration request aborted");
+                return;
+            }
+            
+            Nozzle nozzle = nozzleTip.getParentNozzle();
+            Camera camera = VisionUtils.getBottomVisionCamera();
+            
             try {
                 calibrating = true;
                 
-                reset();
-
-                Nozzle nozzle = nozzleTip.getParentNozzle();
-                Camera camera = VisionUtils.getBottomVisionCamera();
+            	reset();
 
                 // Move to the camera with an angle of 0.
-                Location location = camera.getLocation();
-                location = location.derive(null, null, null, 0d);
-                MovableUtils.moveToLocationAtSafeZ(nozzle, location);
-                for (int i = 0; i < 3; i++) {
-                    // Locate the nozzle offsets.
-                    Location offset = findCircle();
+                Location cameraLocation = camera.getLocation();
+                // This is our baseline location
+                Location measureBaseLocation = cameraLocation.derive(null, null, null, 0d);
 
-                    // Subtract the offsets and move to that position to center the nozzle.
-                    location = location.subtract(offset);
-                    nozzle.moveTo(location);
-                }
-                // This is our baseline location and should have the nozzle well centered over the
-                // camera.
-                Location startLocation = location;
-
-                // Now we rotate the nozzle 360 degrees at calibration.angleIncrement steps, find the
-                // nozzle using the camera and record the offsets.
-                List<CalibrationOffset> offsets = new ArrayList<>();
-                for (double i = 0; i < 360; i += angleIncrement) {
-                    location = startLocation.derive(null, null, null, i);
-                    nozzle.moveTo(location);
-                    Location offset = findCircle();
-                    offsets.add(new CalibrationOffset(offset, i));
-                }
-
-                // The nozzle tip is now calibrated and calibration.getCalibratedOffset() can be
-                // used.
-                this.offsets = offsets;
+                HashMap<String, Object> params = new HashMap<>();
+                params.put("nozzle", nozzle);
+                params.put("camera", camera);
+                Configuration.get().getScripting().on("NozzleCalibration.Starting", params);
                 
-                nozzle.moveToSafeZ();
+                // move nozzle to the camera location at zero degree - the nozzle must not necessarily be at the center
+                MovableUtils.moveToLocationAtSafeZ(nozzle, measureBaseLocation);
+
+                // determine the resulting angleIncrements
+                double angleIncrement = ( angleStop - angleStart ) / this.angleSubdivisions;
+                
+                // determine the number of measurements to be made
+                int angleSubdivisions = this.angleSubdivisions;
+                if(angleStart == -180 && angleStop == 180) {
+                    // on a normal machine start is at 0°, stop would be at 360°. since the nozzle tip is at the same position then, the last measurement can be omitted
+                    angleSubdivisions--;
+                }
+                
+                Logger.debug("[nozzleTipCalibration]starting measurement; angleStart: {}, angleStop: {}, angleIncrement: {}, angleSubdivisions: {}", angleStart, angleStop, angleIncrement, angleSubdivisions);
+                
+                // Capture nozzle tip positions and add them to a list. For these calcs the camera location is considered to be 0/0
+                List<Location> nozzleTipMeasuredLocations = new ArrayList<>();
+                for (int i = 0; i <= angleSubdivisions; i++) {
+                    // calc the current measurement-angle
+                    double measureAngle = angleStart + (i * angleIncrement); 
+                    
+                    Logger.debug("[nozzleTipCalibration]i: {}, measureAngle: {}", i, measureAngle);
+                    
+                	// rotate nozzle to measurement angle
+                    Location measureLocation = measureBaseLocation.derive(null, null, null, measureAngle);
+                    nozzle.moveTo(measureLocation);
+                    
+                    // detect the nozzle tip
+                    Location offset = findCircle();
+                    offset = offset.derive(null, null, null, measureAngle);		// for later usage in the algorithm, the measureAngle is stored to the offset location 
+                    
+                    // add offset to array
+                    nozzleTipMeasuredLocations.add(offset);
+                    
+                    Logger.trace("[nozzleTipCalibration]measured offset: {}", offset);
+                    
+                }
+                
+                Configuration.get().getScripting().on("NozzleCalibration.Finished", params);
+                
+            	if (this.runoutCompensationAlgorithm == RunoutCompensationAlgorithm.Model) {
+            	    this.runoutCompensation = new ModelBasedRunoutCompensation(nozzleTipMeasuredLocations);
+            	} else {
+            	    this.runoutCompensation = new TableBasedRunoutCompensation(nozzleTipMeasuredLocations);
+            	}
+                
             }
             finally {
+                // go to camera position (now offset-corrected). prevents the user from being irritated if it's not exactly centered
+                nozzle.moveTo(camera.getLocation().derive(null, null, null, 0d));
+                
+                // after processing the nozzle returns to safe-z
+                nozzle.moveToSafeZ();
+                
+                // setting to false in the very end to prevent endless calibration repetitions if calibration was not successful (pipeline not well or similar) and the nozzle is commanded afterwards somewhere else (where the calibration is asked for again ...)
                 calibrating = false;
+                
+                // inform UI about new information available
+                firePropertyChange("runoutCompensationInformation", null, null);
             }
         }
-
+        
+        /*
+         * While calibrating the nozzle a circle was fitted to the runout path of the tip.
+         * here the offset is reconstructed in XY-cartesian coordinates to be applied in moveTo commands.
+         */
         public Location getCalibratedOffset(double angle) {
             if (!isEnabled() || !isCalibrated()) {
                 return new Location(LengthUnit.Millimeters, 0, 0, 0, 0);
             }
 
-            // Make sure the angle is between 0 and 360.
-            while (angle < 0) {
+            // Make sure the angle is between -180 and 180 - angles can get larger/smaller as +-180 if limitation to 180 degrees is disabled
+            while (angle < 180) {
                 angle += 360;
             }
-            while (angle > 360) {
+            while (angle > 180) {
                 angle -= 360;
             }
-            List<CalibrationOffset> offsets = getOffsetPairForAngle(angle);
-            CalibrationOffset a = offsets.get(0);
-            CalibrationOffset b = offsets.get(1);
-            Location offsetA = a.offset;
-            Location offsetB = b.offset.convertToUnits(a.offset.getUnits());
-
-            double ratio = (angle - a.angle) / (b.angle - a.angle);
-            double deltaX = offsetB.getX() - offsetA.getX();
-            double deltaY = offsetB.getY() - offsetA.getY();
-            double offsetX = offsetA.getX() + (deltaX * ratio);
-            double offsetY = offsetA.getY() + (deltaY * ratio);
-
-            return new Location(offsetA.getUnits(), offsetX, offsetY, 0, 0);
+            
+            return this.runoutCompensation.getOffset(angle);
+            
         }
 
         private Location findCircle() throws Exception {
@@ -468,20 +775,18 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
                 pipeline.setProperty("camera", camera);
                 pipeline.process();
                 Location location;
+                
                 Object result = pipeline.getResult(VisionUtils.PIPELINE_RESULTS_NAME).model;
+                
+                // are there any results from the pipeline?
+                if (0==((List) result).size()) {
+                    throw new Exception("No results from vision. Check pipeline.");                    
+                }
+
                 if (result instanceof List) {
                     if (((List) result).get(0) instanceof Result.Circle) {
-                        List<Result.Circle> circles = (List<Result.Circle>) result;
-                        List<Location> locations = circles.stream().map(circle -> {
-                            return VisionUtils.getPixelCenterOffsets(camera, circle.x, circle.y);
-                        }).sorted((a, b) -> {
-                            double a1 =
-                                    a.getLinearDistanceTo(new Location(LengthUnit.Millimeters, 0, 0, 0, 0));
-                            double b1 =
-                                    b.getLinearDistanceTo(new Location(LengthUnit.Millimeters, 0, 0, 0, 0));
-                            return Double.compare(a1, b1);
-                        }).collect(Collectors.toList());
-                        location = locations.get(0);
+                        Result.Circle circle = ((List<Result.Circle>) result).get(0);
+                        location = VisionUtils.getPixelCenterOffsets(camera, circle.x, circle.y);
                     }
                     else if (((List) result).get(0) instanceof KeyPoint) {
                         KeyPoint keyPoint = ((List<KeyPoint>) result).get(0);
@@ -491,36 +796,16 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
                         throw new Exception("Unrecognized result " + result);
                     }
                 }
-                else if (result instanceof RotatedRect) {
-                    RotatedRect rect = (RotatedRect) result;
-                    location = VisionUtils.getPixelCenterOffsets(camera, rect.center.x, rect.center.y);
-                }
                 else {
                     throw new Exception("Unrecognized result " + result);
                 }
+
+                //show result from pipeline in camera view
                 MainFrame.get().get().getCameraViews().getCameraView(camera).showFilteredImage(
-                        OpenCvUtils.toBufferedImage(pipeline.getWorkingImage()), 250);
+                        OpenCvUtils.toBufferedImage(pipeline.getWorkingImage()), 1000);
+                
                 return location;
             }
-        }
-
-        /**
-         * Find the two closest offsets to the angle being requested. The offsets start at angle 0
-         * and go to angle 360 - angleIncrement in angleIncrement steps.
-         */
-        private List<CalibrationOffset> getOffsetPairForAngle(double angle) {
-            CalibrationOffset a = null, b = null;
-            if (angle >= offsets.get(offsets.size() - 1).angle) {
-                return Arrays.asList(offsets.get(offsets.size() - 1), offsets.get(0));
-            }
-            for (int i = 0; i < offsets.size(); i++) {
-                if (angle < offsets.get(i + 1).angle) {
-                    a = offsets.get(i);
-                    b = offsets.get(i + 1);
-                    break;
-                }
-            }
-            return Arrays.asList(a, b);
         }
 
         public static CvPipeline createDefaultPipeline() {
@@ -534,26 +819,37 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
             }
         }
 
+        public void resetPipeline() {
+            pipeline = createDefaultPipeline();
+        }
+
         public void reset() {
-            offsets = null;
+        	runoutCompensation = null;
+        	
+        	// inform UI about removed information
+            firePropertyChange("runoutCompensationInformation", null, null);
         }
 
         public boolean isCalibrated() {
-            return offsets != null && !offsets.isEmpty();
+            return runoutCompensation != null;
         }
         
         public boolean isCalibrating() {
             return calibrating;
         }
         
+        public int getAngleSubdivisions() {
+            return angleSubdivisions;
+        }
+
+        public void setAngleSubdivisions(int angleSubdivisions) {
+            this.angleSubdivisions = angleSubdivisions;
+        }
+
         public boolean isEnabled() {
             return enabled;
         }
         
-        public boolean isCalibrationNeeded() {
-            return isEnabled() && !isCalibrated() && !isCalibrating();
-        }
-
         public void setEnabled(boolean enabled) {
             this.enabled = enabled;
         }